+++ settings.lua
@@ -1,7 +1,7 @@
 -- ## Init Settings
 Settings = {}
 
-Settings.difficulty = 3
-Settings.match_win = 3
+Settings.difficulty = 2
+Settings.match_win = 11
 
 return Settings